requirements.txt
Pillow
reportlab
diffusers
accelerate
transformers
google-generativeai
markdown
weasyprint
PyPDF2
peft
torch
python-dotenv
streamlit